+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/WebviewMessage.kt
@@ -46,6 +46,7 @@ sealed class WebviewMessage {
           "reset" -> context.deserialize<ResetWebviewMessage>(element, ResetWebviewMessage::class.java)
           "attribution-search" -> context.deserialize<`attribution-searchWebviewMessage`>(element, `attribution-searchWebviewMessage`::class.java)
           "troubleshoot/reloadAuth" -> context.deserialize<Troubleshoot_reloadAuthWebviewMessage>(element, Troubleshoot_reloadAuthWebviewMessage::class.java)
+          "getAllMentionProvidersMetadata" -> context.deserialize<GetAllMentionProvidersMetadataWebviewMessage>(element, GetAllMentionProvidersMetadataWebviewMessage::class.java)
           else -> throw Exception("Unknown discriminator ${element}")
         }
       }
@@ -410,6 +411,15 @@ data class Troubleshoot_reloadAuthWebviewMessage(
 
   enum class CommandEnum {
     @SerializedName("troubleshoot/reloadAuth") Troubleshoot_reloadAuth,
+  }
+}
+
+data class GetAllMentionProvidersMetadataWebviewMessage(
+  val command: CommandEnum, // Oneof: getAllMentionProvidersMetadata
+) : WebviewMessage() {
+
+  enum class CommandEnum {
+    @SerializedName("getAllMentionProvidersMetadata") GetAllMentionProvidersMetadata,
   }
 }