+++ lib/shared/src/context/openctx/api.ts
@@ -1,23 +1,21 @@
 import type { Client } from '@openctx/client'
 import type * as vscode from 'vscode'
 
-type OpenCtxClient = Client<vscode.Range>
+export type OpenCtxClient = Client<vscode.Range>
 
-let _client: OpenCtxClient | undefined
+class OpenCtx {
+    constructor(public client: OpenCtxClient | undefined) {}
+}
+
+export const openCtx = new OpenCtx(undefined)
 
 /**
  * Set the handle to the OpenCtx client.
  */
-export function setOpenCtxClient(client: OpenCtxClient | undefined): void {
-    if (_client) {
+export function setOpenCtxClient(client: OpenCtxClient): void {
+    if (openCtx.client) {
         throw new Error('OpenCtx extension API is already set')
     }
-    _client = client
-}
 
-/**
- * Get a handle to the OpenCtx client, set in {@link setOpenCtxClient}.
- */
-export function getOpenCtxClient(): OpenCtxClient | undefined {
-    return _client
+    openCtx.client = client
 }